melhorias do código.

+++ Gp.Domain/Models/ReceitaLancamento.cs
@@ -3,7 +3,7 @@
     public class ReceitaLancamento : Base
     {
         public string Descricao { get; set; }
-        public float ValorTotal { get; set; }
+        public decimal ValorTotal { get; set; }
         public int ReceitaId { get; set; }
         public virtual Receita Receita { get; set; }
     }
